cube 2 backports, servcmd tweaks for echoing current state to all players, hud (over|under)lay textures
Former-commit-id: 396a693d7e1fe1178349d947b7e2717e03b1a70d

+++ src/shared/stream.cpp
@@ -332,7 +332,7 @@ SDL_RWops *stream::rwops()
 
 long stream::size()
 {
-    int pos = tell(), endpos;
+    long pos = tell(), endpos;
     if(pos < 0 || !seek(0, SEEK_END)) return -1;
     endpos = tell();
     return pos == endpos || seek(pos, SEEK_SET) ? endpos : -1;
